Rename WooPosDesignSystemUsage to WooPosDesignSystemUsageSpacing and update spacing validation logic
