+++ libs/detektrules/src/main/resources/config/config.yml
@@ -1,3 +1,3 @@
 custom-rules:
-  WooPosDesignSystemUsage:
+  WooPosDesignSystemUsageSpacing:
     active: true
